search bar fully functional on all products page
