+++ client/routes.js
@@ -37,10 +37,6 @@ class Routes extends Component {
         {/* Routes placed here are available to all visitors */}
         <Route path="/login" component={Login} />
         <Route path="/signup" component={Signup} />
-        <Route
-          path="/products/searchResults/:searchTerms"
-          component={productSearchResults}
-        />
         <Route path="/products/:productId" component={SingleProduct} />
         <Route
           path="/products"
